Tiny optimization for copy_vm_xcp / backup_xva schemes
Both methods create a snap and make a copy of it, and both methods
ought to be used in the same backup process.  To avoid overwhelming
the storage, it is now possible to avoid deleting a snap in
copy_vm_xcp (delete_snapshot=no) so that it gets reused in
backup_xva (reuse_snapshot=yes).

Requested by Yvan.